requirements.txt
blinker==1.9.0
click==8.3.1
flask==3.1.2
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
python-dotenv==1.2.1
werkzeug==3.1.5
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
distro==1.9.0
executing==2.2.1
fastapi==0.128.0
frozenlist==1.8.0
google-auth==2.45.0
google-genai==1.56.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
icecream==2.1.8
idna==3.11
iniconfig==2.3.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-text-splitters==1.1.0
langsmith==0.5.2
markdown-it-py==4.0.0
marshmallow==3.26.2
mcp==1.25.0
mcp-kipris @ git+https://github.com/nuri428/mcp_kipris.git@e404958f1b9e16778675f0f98fb49609d96abc71
mdurl==0.1.2
multidict==6.7.0
mypy-extensions==1.1.0
networkx==3.6.1
numpy==2.4.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
propcache==0.4.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
pyyaml==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.16.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.45
sse-starlette==3.1.1
starlette==0.50.0
stringcase==1.2.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
typer==0.21.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.2
uuid-utils==0.12.0
uvicorn==0.40.0
websockets==15.0.1
xmltodict==1.0.2
yarl==1.22.0
zstandard==0.25.0
